+++ src/runtime/extended-access/gaa-google-sign-in-button-test.js
@@ -1,0 +1,293 @@
+/**
+ * Copyright 2018 The Subscribe with Google Authors. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS-IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import {GOOGLE_SIGN_IN_BUTTON_ID} from './html-templates';
+import {GaaGoogleSignInButton} from './';
+import {I18N_STRINGS} from '../../i18n/strings';
+import {
+  POST_MESSAGE_COMMAND_ERROR,
+  POST_MESSAGE_COMMAND_GSI_BUTTON_CLICK,
+  POST_MESSAGE_COMMAND_INTRODUCTION,
+  POST_MESSAGE_COMMAND_USER,
+  POST_MESSAGE_STAMP,
+} from './constants';
+import {tick} from '../../../test/tick';
+import {QueryStringUtils} from './utils';
+
+describes.realWin('GaaGoogleSignInButton', () => {
+  const allowedOrigins = [location.origin];
+
+  let clock;
+
+  beforeEach(() => {
+    // Mock clock.
+    clock = sandbox.useFakeTimers();
+
+    self.gapi = {
+      load: sandbox.fake((name, callback) => {
+        callback();
+      }),
+      auth2: {
+        init: sandbox.fake(),
+        getAuthInstance: sandbox.fake(),
+      },
+      signin2: {
+        render: sandbox.fake(),
+      },
+    };
+
+    // Mock query string.
+    sandbox.stub(QueryStringUtils, 'getQueryString');
+    QueryStringUtils.getQueryString.returns('?lang=en');
+
+    // Mock console.warn method.
+    sandbox.stub(self.console, 'warn');
+  });
+
+  afterEach(() => {
+    if (self.postMessage.restore) {
+      self.postMessage.restore();
+    }
+    QueryStringUtils.getQueryString.restore();
+
+    // Remove the injected style from GaaGoogleSignInButton.show.
+    for (const style of [...self.document.head.querySelectorAll('style')]) {
+      style.remove();
+    }
+
+    self.console.warn.restore();
+  });
+
+  describe('show', () => {
+    it('renders Google Sign-In button', async () => {
+      GaaGoogleSignInButton.show({allowedOrigins});
+      clock.tick(100);
+      await tick(10);
+
+      const args = self.gapi.signin2.render.args;
+      expect(typeof args[0][1].onsuccess).to.equal('function');
+      expect(args).to.deep.equal([
+        [
+          'swg-google-sign-in-button',
+          {
+            longtitle: true,
+            onsuccess: args[0][1].onsuccess,
+            prompt: 'select_account',
+            scope: 'profile email',
+            theme: 'dark',
+          },
+        ],
+      ]);
+    });
+
+    it('renders supported i18n languages', async () => {
+      QueryStringUtils.getQueryString.returns('?lang=pt-br');
+
+      GaaGoogleSignInButton.show({allowedOrigins});
+      clock.tick(100);
+      await tick(10);
+
+      const styleEl = self.document.querySelector('style');
+      expect(styleEl.textContent).to.contain(
+        I18N_STRINGS.SHOWCASE_REGWALL_GOOGLE_SIGN_IN_BUTTON['pt-br']
+      );
+    });
+
+    it('renders English by default, if "lang" URL param is missing', async () => {
+      QueryStringUtils.getQueryString.returns('?');
+
+      GaaGoogleSignInButton.show({allowedOrigins});
+      clock.tick(100);
+      await tick(10);
+
+      const styleEl = self.document.querySelector('style');
+      expect(styleEl.textContent).to.contain(
+        I18N_STRINGS.SHOWCASE_REGWALL_GOOGLE_SIGN_IN_BUTTON['en']
+      );
+    });
+
+    it('sends post message with button click event', async () => {
+      // Show button.
+      GaaGoogleSignInButton.show({allowedOrigins});
+
+      // Send intro post message.
+      postMessage({
+        stamp: POST_MESSAGE_STAMP,
+        command: POST_MESSAGE_COMMAND_INTRODUCTION,
+      });
+
+      // Wait for promises and intervals to resolve.
+      clock.tick(100);
+      await tick(10);
+
+      // Click button.
+      self.document.getElementById(GOOGLE_SIGN_IN_BUTTON_ID).click();
+
+      // Wait for button click post message.
+      await new Promise((resolve) => {
+        sandbox.stub(self, 'postMessage').callsFake(() => {
+          resolve();
+        });
+      });
+
+      // Expect button click post message.
+      expect(self.postMessage).to.be.calledWithExactly(
+        {
+          command: POST_MESSAGE_COMMAND_GSI_BUTTON_CLICK,
+          stamp: POST_MESSAGE_STAMP,
+        },
+        location.origin
+      );
+    });
+
+    it('sends post message with GAA user', async () => {
+      GaaGoogleSignInButton.show({allowedOrigins});
+
+      // Send intro post message.
+      postMessage({
+        stamp: POST_MESSAGE_STAMP,
+        command: POST_MESSAGE_COMMAND_INTRODUCTION,
+      });
+
+      // Mock Google Sign-In response with GoogleUser object.
+      const gaaUser = {
+        idToken: 'idToken',
+        name: 'name',
+        givenName: 'givenName',
+        familyName: 'familyName',
+        imageUrl: 'imageUrl',
+        email: 'email',
+        authorizationData: {
+          /* eslint-disable google-camelcase/google-camelcase */
+          access_token: 'accessToken',
+          id_token: 'idToken',
+          scope: 'scope',
+          expires_in: 0,
+          first_issued_at: 0,
+          expires_at: 0,
+          /* eslint-enable google-camelcase/google-camelcase */
+        },
+      };
+      const googleUser = {
+        getBasicProfile: () => ({
+          getName: () => gaaUser.name,
+          getGivenName: () => gaaUser.givenName,
+          getFamilyName: () => gaaUser.familyName,
+          getImageUrl: () => gaaUser.imageUrl,
+          getEmail: () => gaaUser.email,
+        }),
+        getAuthResponse: () => gaaUser.authorizationData,
+      };
+      const args = self.gapi.signin2.render.args;
+      // Wait for promises and intervals to resolve.
+      clock.tick(100);
+      await tick(10);
+      // Send GoogleUser.
+      args[0][1].onsuccess(googleUser);
+
+      // Wait for post message.
+      await new Promise((resolve) => {
+        sandbox.stub(self, 'postMessage').callsFake(() => {
+          resolve();
+        });
+      });
+
+      expect(self.postMessage).to.be.calledWithExactly(
+        {
+          command: POST_MESSAGE_COMMAND_USER,
+          gaaUser: {
+            email: 'email',
+            familyName: 'familyName',
+            givenName: 'givenName',
+            idToken: 'idToken',
+            imageUrl: 'imageUrl',
+            name: 'name',
+            authorizationData: {
+              /* eslint-disable google-camelcase/google-camelcase */
+              access_token: 'accessToken',
+              id_token: 'idToken',
+              scope: 'scope',
+              expires_in: 0,
+              first_issued_at: 0,
+              expires_at: 0,
+              /* eslint-enable google-camelcase/google-camelcase */
+            },
+          },
+          stamp: POST_MESSAGE_STAMP,
+        },
+        location.origin
+      );
+    });
+
+    it('sends errors to parent', async () => {
+      self.gapi.signin2.render = sandbox.fake.throws('I need cookies');
+
+      GaaGoogleSignInButton.show({allowedOrigins});
+
+      // Send intro post message.
+      postMessage({
+        stamp: POST_MESSAGE_STAMP,
+        command: POST_MESSAGE_COMMAND_INTRODUCTION,
+      });
+
+      // Wait for promises and intervals to resolve.
+      clock.tick(100);
+      await tick(10);
+
+      // Wait for post message.
+      await new Promise((resolve) => {
+        sandbox.stub(self, 'postMessage').callsFake(() => {
+          resolve();
+        });
+      });
+
+      expect(self.postMessage).to.be.calledWithExactly(
+        {
+          command: POST_MESSAGE_COMMAND_ERROR,
+          stamp: POST_MESSAGE_STAMP,
+        },
+        location.origin
+      );
+    });
+
+    it('fails and warns when passed invalid origins', async () => {
+      const invalidOrigins = [
+        // Bad protocol, should be http or https.
+        'ftp://localhost:8080',
+        // Includes path.
+        'http://localhost:8080/',
+      ];
+
+      for (const invalidOrigin of invalidOrigins) {
+        GaaGoogleSignInButton.show({allowedOrigins: [invalidOrigin]});
+
+        // Send intro post message.
+        postMessage({
+          stamp: POST_MESSAGE_STAMP,
+          command: POST_MESSAGE_COMMAND_INTRODUCTION,
+        });
+
+        // Wait for promises and intervals to resolve.
+        clock.tick(100);
+        await tick(10);
+
+        expect(self.console.warn).to.have.been.calledWithExactly(
+          `[swg-gaa.js:GaaGoogleSignInButton.show]: You specified an invalid origin: ${invalidOrigin}`
+        );
+      }
+    });
+  });
+});
