+++ src/runtime/extended-access/gaa-metering-test.js
@@ -15,42 +15,13 @@
  */
 
 import {AnalyticsEvent} from '../../proto/api_messages';
-import {
-  GOOGLE_3P_SIGN_IN_BUTTON_ID,
-  GOOGLE_SIGN_IN_BUTTON_ID,
-  GOOGLE_SIGN_IN_IFRAME_ID,
-  REGWALL_CONTAINER_ID,
-  REGWALL_DIALOG_ID,
-  REGWALL_TITLE_ID,
-  SIGN_IN_WITH_GOOGLE_BUTTON_ID,
-} from './html-templates';
-import {
-  GaaGoogle3pSignInButton,
-  GaaGoogleSignInButton,
-  GaaMetering,
-  GaaMeteringRegwall,
-  GaaSignInWithGoogleButton,
-  queryStringHasFreshGaaParams,
-} from './';
-import {I18N_STRINGS} from '../../i18n/strings';
-import {JwtHelper} from '../../utils/jwt';
-import {
-  POST_MESSAGE_COMMAND_3P_BUTTON_CLICK,
-  POST_MESSAGE_COMMAND_ERROR,
-  POST_MESSAGE_COMMAND_GSI_BUTTON_CLICK,
-  POST_MESSAGE_COMMAND_INTRODUCTION,
-  POST_MESSAGE_COMMAND_SIWG_BUTTON_CLICK,
-  POST_MESSAGE_COMMAND_USER,
-  POST_MESSAGE_STAMP,
-} from './constants';
+import {GaaMetering, GaaMeteringRegwall} from '.';
 import {ShowcaseEvent} from '../../api/subscriptions';
 import {tick} from '../../../test/tick';
 import {QueryStringUtils} from './utils';
 
 const PUBLISHER_NAME = 'The Scenic';
 const PRODUCT_ID = 'scenic-2017.appspot.com:news';
-const GSI_IFRAME_URL = 'https://localhost/gsi-iframe';
-const SIWG_IFRAME_URL = 'https://localhost/gis-iframe';
 const GOOGLE_3P_AUTH_URL = 'https://fabulous-3p-authserver.glitch.me/auth';
 const CASL_URL = 'https://example-casl.com';
 const GOOGLE_API_CLIENT_ID = 'test123.apps.googleusercontent.com';
@@ -173,2048 +144,6 @@ const SIGN_IN_WITH_GOOGLE_DECODED_JWT = {
     /* eslint-enable google-camelcase/google-camelcase */
   },
 };
-
-describes.realWin('queryStringHasFreshGaaParams', () => {
-  let clock;
-
-  beforeEach(() => {
-    clock = sandbox.useFakeTimers();
-  });
-
-  it('succeeeds for valid params', () => {
-    const queryString = '?gaa_at=at&gaa_n=n&gaa_sig=sig&gaa_ts=99999';
-    expect(queryStringHasFreshGaaParams(queryString)).to.be.true;
-  });
-
-  it('fails without gaa_at', () => {
-    const queryString = '?gaa_n=n0nc3&gaa_sig=s1gn4tur3&gaa_ts=99999';
-    expect(queryStringHasFreshGaaParams(queryString)).to.be.false;
-  });
-
-  it('fails without gaa_n', () => {
-    const queryString = '?gaa_at=gaa&gaa_sig=s1gn4tur3&gaa_ts=99999';
-    expect(queryStringHasFreshGaaParams(queryString)).to.be.false;
-  });
-
-  it('fails without gaa_sig', () => {
-    const queryString = '?gaa_at=gaa&gaa_n=n0nc3&gaa_ts=99999';
-    expect(queryStringHasFreshGaaParams(queryString)).to.be.false;
-  });
-
-  it('fails without gaa_ts', () => {
-    const queryString = '?gaa_at=gaa&gaa_n=n0nc3&gaa_sig=s1gn4tur3';
-    expect(queryStringHasFreshGaaParams(queryString)).to.be.false;
-  });
-
-  it('fails if GAA URL params are expired', () => {
-    // Add GAA URL params with expiration of 7 seconds.
-    const queryString = '?gaa_at=gaa&gaa_n=n0nc3&gaa_sig=s1gn4tur3&gaa_ts=7';
-    clock.tick(7001);
-    expect(queryStringHasFreshGaaParams(queryString)).to.be.false;
-  });
-
-  it('fails if gaa_at param specifies "no access"', () => {
-    const queryString = '?gaa_at=na&gaa_n=n&gaa_sig=sig&gaa_ts=99999';
-    expect(queryStringHasFreshGaaParams(queryString)).to.be.false;
-  });
-
-  it('succeeds if gaa_at param specifies "no access" but allowAllAccessTypes is true', () => {
-    const queryString = '?gaa_at=na&gaa_n=n&gaa_sig=sig&gaa_ts=99999';
-    expect(queryStringHasFreshGaaParams(queryString, true)).to.be.true;
-  });
-});
-
-describes.realWin('GaaMeteringRegwall', () => {
-  let clock;
-  let logEvent;
-  let microdata;
-  let script;
-  let signOutFake;
-  let subscriptionsMock;
-  let renderSpy;
-
-  beforeEach(() => {
-    // Mock clock.
-    clock = sandbox.useFakeTimers();
-
-    // Mock Google Sign-In API.
-    signOutFake = sandbox.fake.resolves();
-    let authInstance;
-    self.gapi = {
-      load: sandbox.fake((name, callback) => {
-        callback();
-      }),
-      auth2: {
-        init: sandbox.fake(() => {
-          authInstance = {
-            signOut: signOutFake,
-          };
-        }),
-        getAuthInstance: sandbox.fake(() => authInstance),
-      },
-      signin2: {
-        render: sandbox.fake(),
-      },
-    };
-
-    // Mock Sign In with Google API.
-    self.google = {
-      accounts: {
-        id: {
-          initialize: sandbox.fake(),
-          renderButton: sandbox.fake(),
-        },
-      },
-    };
-
-    // Mock SwG API.
-    logEvent = sandbox.fake();
-    subscriptionsMock = {
-      triggerLoginRequest: sandbox.fake(),
-      getEventManager: () => Promise.resolve({logEvent}),
-    };
-    self.SWG = {
-      push: sandbox.fake((callback) => void callback(subscriptionsMock)),
-    };
-
-    // Mock query string.
-    sandbox.stub(QueryStringUtils, 'getQueryString');
-    QueryStringUtils.getQueryString.returns(
-      '?gaa_at=gaa&gaa_n=n0nc3&gaa_sig=s1gn4tur3&gaa_ts=99999999'
-    );
-
-    // Mock window opener
-    sandbox.stub(self, 'open');
-
-    // Mock console.warn & log methods.
-    sandbox.stub(self.console, 'warn');
-    sandbox.stub(self.console, 'log');
-
-    // Add JSON-LD with a publisher name.
-    script = self.document.createElement('script');
-    script.type = 'application/ld+json';
-    script.text = ARTICLE_LD_JSON_METADATA;
-    self.document.head.appendChild(script);
-
-    // Add container for Microdata.
-    microdata = self.document.createElement('div');
-    self.document.head.appendChild(microdata);
-  });
-
-  afterEach(() => {
-    script.remove();
-    microdata.remove();
-    GaaMeteringRegwall.remove();
-    self.document.documentElement.lang = '';
-    // Remove the injected style from GaaMeteringRegwall.createNativeRegistrationButton.
-    self.document.head.querySelectorAll('style').forEach((e) => {
-      e.remove();
-    });
-
-    self.console.warn.restore();
-    self.console.log.restore();
-  });
-
-  /**
-   * Expects a list of Analytics events.
-   * @param {!Array<{
-   *   analyticsEvent: !ShowcaseEvent,
-   *   isFromUserAction: boolean,
-   * }>} events
-   */
-  function expectAnalyticsEvents(events) {
-    expect(logEvent).to.have.callCount(events.length);
-    for (let i = 0; i < events.length; i++) {
-      const {analyticsEvent, isFromUserAction} = events[i];
-      expect(logEvent.getCall(i)).to.be.calledWithExactly({
-        eventType: analyticsEvent,
-        eventOriginator: 1,
-        isFromUserAction,
-        additionalParameters: null,
-      });
-    }
-  }
-
-  describe('render_', () => {
-    it('shows regwall with publisher name', () => {
-      GaaMeteringRegwall.render_({iframeUrl: GSI_IFRAME_URL});
-
-      const descriptionEl = self.document.querySelector(
-        '.gaa-metering-regwall--description'
-      );
-      expect(descriptionEl.textContent).contains(PUBLISHER_NAME);
-    });
-
-    it('does not render CASL blurb by default', () => {
-      GaaMeteringRegwall.render_({iframeUrl: GSI_IFRAME_URL});
-
-      const caslEl = self.document.querySelector('.gaa-metering-regwall--casl');
-      expect(caslEl).to.be.null;
-    });
-
-    it('optionally renders CASL blurb', () => {
-      GaaMeteringRegwall.render_({
-        iframeUrl: GSI_IFRAME_URL,
-        caslUrl: CASL_URL,
-      });
-
-      const caslEl = self.document.querySelector('.gaa-metering-regwall--casl');
-      expect(caslEl.textContent).contains("Review The Scenic's CASL terms");
-
-      const caslLinkEl = caslEl.querySelector('a');
-      expect(caslLinkEl.href).contains(CASL_URL);
-    });
-
-    it('focuses on modal title after the animation completes', () => {
-      GaaMeteringRegwall.render_({iframeUrl: GSI_IFRAME_URL});
-
-      // Mock animation ending.
-      const dialogEl = self.document.getElementById(REGWALL_DIALOG_ID);
-      dialogEl.dispatchEvent(new Event('animationend'));
-
-      const titleEl = self.document.getElementById(REGWALL_TITLE_ID);
-      expect(self.document.activeElement).to.equal(titleEl);
-    });
-
-    it('parses publisher name from microdata', () => {
-      // Remove JSON-LD.
-      script.text = '{}';
-
-      // Add Microdata.
-      microdata.innerHTML = ARTICLE_MICRODATA_METADATA;
-
-      GaaMeteringRegwall.render_({iframeUrl: GSI_IFRAME_URL});
-
-      const descriptionEl = self.document.querySelector(
-        '.gaa-metering-regwall--description'
-      );
-      expect(descriptionEl.textContent).contains(PUBLISHER_NAME);
-    });
-
-    it('throws if article metadata lacks a publisher name', () => {
-      // Remove JSON-LD.
-      script.text = '{}';
-
-      const showingRegwall = () =>
-        GaaMeteringRegwall.render_({iframeUrl: GSI_IFRAME_URL});
-
-      expect(showingRegwall).throws(
-        'Showcase articles must define a publisher name with either JSON-LD or Microdata.'
-      );
-    });
-
-    it('renders supported i18n languages', () => {
-      self.document.documentElement.lang = 'pt-br';
-
-      GaaMeteringRegwall.render_({iframeUrl: GSI_IFRAME_URL});
-
-      const titleEl = self.document.querySelector(
-        '.gaa-metering-regwall--title'
-      );
-      expect(titleEl.textContent).to.equal(
-        I18N_STRINGS.SHOWCASE_REGWALL_TITLE['pt-br']
-      );
-    });
-
-    it('renders "en" for non-supported i18n languages', () => {
-      self.document.documentElement.lang = 'non-supported';
-
-      GaaMeteringRegwall.render_({iframeUrl: GSI_IFRAME_URL});
-
-      const titleEl = self.document.querySelector(
-        '.gaa-metering-regwall--title'
-      );
-      expect(titleEl.textContent).to.equal(
-        I18N_STRINGS.SHOWCASE_REGWALL_TITLE['en']
-      );
-    });
-
-    it('adds "lang" URL param to iframe URL', () => {
-      self.document.documentElement.lang = 'pt-br';
-
-      GaaMeteringRegwall.render_({iframeUrl: GSI_IFRAME_URL});
-
-      const iframeEl = self.document.getElementById(GOOGLE_SIGN_IN_IFRAME_ID);
-      expect(iframeEl.src).to.contain('?lang=pt-br');
-    });
-
-    it('handles clicks on publisher sign in link', async () => {
-      // Show Regwall.
-      GaaMeteringRegwall.render_({iframeUrl: GSI_IFRAME_URL});
-      await tick();
-      logEvent.resetHistory();
-
-      // Click publisher link to trigger a login request.
-      const publisherSignInButtonEl = self.document.querySelector(
-        '#swg-publisher-sign-in-button'
-      );
-      publisherSignInButtonEl.click();
-      expect(subscriptionsMock.triggerLoginRequest).to.be.calledWithExactly({
-        linkRequested: false,
-      });
-      await tick();
-
-      // Verify analytics event.
-      expectAnalyticsEvents([
-        {
-          analyticsEvent:
-            AnalyticsEvent.ACTION_SHOWCASE_REGWALL_EXISTING_ACCOUNT_CLICK,
-          isFromUserAction: true,
-        },
-      ]);
-    });
-  });
-
-  describe('show', () => {
-    beforeEach(() => {
-      renderSpy = sandbox.spy(GaaMeteringRegwall, 'render_');
-    });
-
-    it('passes iframeUrl to GaaMeteringRegwall.render_', () => {
-      GaaMeteringRegwall.show({iframeUrl: GSI_IFRAME_URL});
-      expect(renderSpy).to.have.been.calledWithExactly({
-        iframeUrl: GSI_IFRAME_URL,
-        caslUrl: undefined,
-      });
-    });
-
-    it('optionally passes caslUrl to GaaMeteringRegwall.render_', () => {
-      GaaMeteringRegwall.show({iframeUrl: GSI_IFRAME_URL, caslUrl: CASL_URL});
-
-      expect(renderSpy).to.have.been.calledWithExactly({
-        iframeUrl: GSI_IFRAME_URL,
-        caslUrl: CASL_URL,
-      });
-    });
-
-    it('returns GAA User', async () => {
-      const gaaUser = {name: 'Hello'};
-      const gaaUserPromise = GaaMeteringRegwall.show({
-        iframeUrl: GSI_IFRAME_URL,
-      });
-
-      postMessage({
-        stamp: POST_MESSAGE_STAMP,
-        command: POST_MESSAGE_COMMAND_USER,
-        gaaUser,
-      });
-
-      expect(await gaaUserPromise).to.deep.equal(gaaUser);
-    });
-
-    it('removes Regwall from DOM', async () => {
-      postMessage({
-        stamp: POST_MESSAGE_STAMP,
-        command: POST_MESSAGE_COMMAND_USER,
-        gaaUser: {},
-      });
-
-      await GaaMeteringRegwall.show({iframeUrl: GSI_IFRAME_URL});
-
-      expect(self.document.getElementById(REGWALL_CONTAINER_ID)).to.be.null;
-    });
-
-    it('fails if GAA URL params are missing', () => {
-      // Remove GAA URL params.
-      QueryStringUtils.getQueryString.restore();
-
-      GaaMeteringRegwall.show({iframeUrl: GSI_IFRAME_URL});
-
-      expect(self.console.warn).to.have.been.calledWithExactly(
-        '[swg-gaa.js:GaaMeteringRegwall.show]: URL needs fresh GAA params.'
-      );
-    });
-
-    it('fails if GAA URL params are expired', () => {
-      // Add GAA URL params with expiration of 7 seconds.
-      QueryStringUtils.getQueryString.returns(
-        '?gaa_at=gaa&gaa_n=n0nc3&gaa_sig=s1gn4tur3&gaa_ts=7'
-      );
-
-      // Move clock a little past 7 seconds.
-      clock.tick(7001);
-
-      GaaMeteringRegwall.show({iframeUrl: GSI_IFRAME_URL});
-
-      expect(self.console.warn).to.have.been.calledWithExactly(
-        '[swg-gaa.js:GaaMeteringRegwall.show]: URL needs fresh GAA params.'
-      );
-    });
-
-    it('handles GSI error', async () => {
-      const gaaUserPromise = GaaMeteringRegwall.show({
-        iframeUrl: GSI_IFRAME_URL,
-      });
-
-      // Send intro post message.
-      postMessage({
-        stamp: POST_MESSAGE_STAMP,
-        command: POST_MESSAGE_COMMAND_ERROR,
-      });
-
-      // Reject promise.
-      await expect(gaaUserPromise).to.eventually.be.rejectedWith(
-        'Google Sign-In could not render'
-      );
-
-      // Remove Regwall from DOM.
-      expect(self.document.getElementById(REGWALL_CONTAINER_ID)).to.be.null;
-    });
-
-    it('logs Showcase impression events', async () => {
-      GaaMeteringRegwall.show({iframeUrl: GSI_IFRAME_URL});
-      await tick();
-
-      // Verify analytics events.
-      expectAnalyticsEvents([
-        {
-          analyticsEvent: AnalyticsEvent.EVENT_NO_ENTITLEMENTS,
-          isFromUserAction: false,
-        },
-        {
-          analyticsEvent: AnalyticsEvent.IMPRESSION_REGWALL,
-          isFromUserAction: false,
-        },
-        {
-          analyticsEvent: AnalyticsEvent.IMPRESSION_SHOWCASE_REGWALL,
-          isFromUserAction: false,
-        },
-      ]);
-    });
-  });
-
-  describe('showWithNativeRegistrationButton', () => {
-    beforeEach(() => {
-      renderSpy = sandbox.spy(GaaMeteringRegwall, 'render_');
-    });
-
-    it('calls GaaMeteringRegwall.render_ with useNativeMode: true', () => {
-      GaaMeteringRegwall.showWithNativeRegistrationButton({
-        googleApiClientId: GOOGLE_API_CLIENT_ID,
-      });
-      expect(renderSpy).to.have.been.calledWithExactly({
-        iframeUrl: '',
-        caslUrl: undefined,
-        useNativeMode: true,
-      });
-    });
-
-    it('optionally passes caslUrl to GaaMeteringRegwall.render_', () => {
-      GaaMeteringRegwall.showWithNativeRegistrationButton({
-        googleApiClientId: GOOGLE_API_CLIENT_ID,
-        caslUrl: CASL_URL,
-      });
-      expect(renderSpy).to.have.been.calledWithExactly({
-        iframeUrl: '',
-        caslUrl: CASL_URL,
-        useNativeMode: true,
-      });
-    });
-
-    it('resolves with a gaaUser removes Regwall from DOM on click', async () => {
-      const gaaUserPromise =
-        GaaMeteringRegwall.showWithNativeRegistrationButton({
-          googleApiClientId: GOOGLE_API_CLIENT_ID,
-        });
-      clock.tick(100);
-
-      // Click button.
-      self.document.getElementById(SIGN_IN_WITH_GOOGLE_BUTTON_ID).click();
-
-      // Simulate the click resolving
-      const args = self.google.accounts.id.initialize.args;
-      args[0][0].callback(SIGN_IN_WITH_GOOGLE_JWT);
-
-      const gaaUser = await gaaUserPromise;
-      expect(gaaUser).to.deep.equal(SIGN_IN_WITH_GOOGLE_JWT);
-      expect(self.document.getElementById(REGWALL_CONTAINER_ID)).to.be.null;
-    });
-
-    it('resolves with a decoded jwt if rawJwt is false', async () => {
-      const gaaUserPromise =
-        GaaMeteringRegwall.showWithNativeRegistrationButton({
-          googleApiClientId: GOOGLE_API_CLIENT_ID,
-          rawJwt: false,
-        });
-
-      // Mock JWT decoding function.
-      sandbox
-        .stub(JwtHelper.prototype, 'decode')
-        .returns(SIGN_IN_WITH_GOOGLE_DECODED_JWT);
-
-      // Click button.
-      self.document.getElementById(SIGN_IN_WITH_GOOGLE_BUTTON_ID).click();
-
-      // Simulate the click resolving
-      const args = self.google.accounts.id.initialize.args;
-      args[0][0].callback(SIGN_IN_WITH_GOOGLE_JWT);
-
-      await tick();
-
-      expect(await gaaUserPromise).to.equal(SIGN_IN_WITH_GOOGLE_DECODED_JWT);
-      expect(self.document.getElementById(REGWALL_CONTAINER_ID)).to.be.null;
-    });
-
-    it('handles Sign In with Google errors', async () => {
-      self.google.accounts.id.initialize = sandbox.fake.throws(
-        'Function not loaded'
-      );
-
-      const gaaUserPromise =
-        GaaMeteringRegwall.showWithNativeRegistrationButton({
-          googleApiClientId: GOOGLE_API_CLIENT_ID,
-        });
-      clock.tick(100);
-      await tick(10);
-
-      // Reject promise.
-      await gaaUserPromise;
-      expect(self.document.getElementById(REGWALL_CONTAINER_ID)).to.be.null;
-    });
-
-    it('logs Sign In with Google errors when in debug mode', async () => {
-      location.hash = '#swg.debug=1';
-
-      self.google.accounts.id.initialize = sandbox.fake.throws(
-        'Function not loaded'
-      );
-
-      const gaaUserPromise =
-        GaaMeteringRegwall.showWithNativeRegistrationButton({
-          googleApiClientId: GOOGLE_API_CLIENT_ID,
-        });
-      clock.tick(100);
-      await tick(10);
-
-      // Reject promise.
-      await gaaUserPromise;
-      expect(self.console.log).to.calledWithExactly(
-        '[Subscriptions]',
-        'Regwall failed: Error: Function not loaded'
-      );
-      expect(self.document.getElementById(REGWALL_CONTAINER_ID)).to.be.null;
-    });
-
-    it('logs Showcase impression events', async () => {
-      GaaMeteringRegwall.showWithNativeRegistrationButton({
-        googleApiClientId: GOOGLE_API_CLIENT_ID,
-      });
-      await tick();
-
-      // Verify analytics events.
-      expectAnalyticsEvents([
-        {
-          analyticsEvent: AnalyticsEvent.EVENT_NO_ENTITLEMENTS,
-          isFromUserAction: false,
-        },
-        {
-          analyticsEvent: AnalyticsEvent.IMPRESSION_REGWALL,
-          isFromUserAction: false,
-        },
-        {
-          analyticsEvent: AnalyticsEvent.IMPRESSION_SHOWCASE_REGWALL,
-          isFromUserAction: false,
-        },
-      ]);
-    });
-
-    it('logs analytics events on button click', async () => {
-      GaaMeteringRegwall.showWithNativeRegistrationButton({
-        googleApiClientId: GOOGLE_API_CLIENT_ID,
-      });
-
-      // Simulate a click event from SIWG.
-      self.google.accounts.id.renderButton.args[0][1].click_listener();
-
-      await tick();
-
-      // Verify analytics events.
-      expectAnalyticsEvents([
-        {
-          analyticsEvent: AnalyticsEvent.EVENT_NO_ENTITLEMENTS,
-          isFromUserAction: false,
-        },
-        {
-          analyticsEvent: AnalyticsEvent.IMPRESSION_REGWALL,
-          isFromUserAction: false,
-        },
-        {
-          analyticsEvent: AnalyticsEvent.IMPRESSION_SHOWCASE_REGWALL,
-          isFromUserAction: false,
-        },
-        {
-          analyticsEvent: AnalyticsEvent.ACTION_SHOWCASE_REGWALL_SIWG_CLICK,
-          isFromUserAction: true,
-        },
-      ]);
-    });
-  });
-
-  describe('showWithNative3PRegistrationButton', () => {
-    let createNative3PRegistrationButtonSpy;
-    beforeEach(() => {
-      renderSpy = sandbox.spy(GaaMeteringRegwall, 'render_');
-      createNative3PRegistrationButtonSpy = sandbox.spy(
-        GaaMeteringRegwall,
-        'createNative3PRegistrationButton'
-      );
-    });
-
-    it('creates a 3P button when authorizationUrl is provided', () => {
-      GaaMeteringRegwall.showWithNative3PRegistrationButton({
-        authorizationUrl: GOOGLE_3P_AUTH_URL,
-      });
-
-      expect(
-        createNative3PRegistrationButtonSpy
-      ).to.have.been.calledWithExactly({
-        authorizationUrl: GOOGLE_3P_AUTH_URL,
-      });
-    });
-
-    it('calls GaaMeteringRegwall.render_ with useNativeMode: true', () => {
-      GaaMeteringRegwall.showWithNative3PRegistrationButton({
-        authorizationUrl: GOOGLE_3P_AUTH_URL,
-      });
-      expect(renderSpy).to.have.been.calledWithExactly({
-        iframeUrl: '',
-        caslUrl: undefined,
-        useNativeMode: true,
-      });
-    });
-
-    it('optionally passes caslUrl to GaaMeteringRegwall.render_', () => {
-      GaaMeteringRegwall.showWithNative3PRegistrationButton({
-        authorizationUrl: GOOGLE_3P_AUTH_URL,
-        caslUrl: CASL_URL,
-      });
-      expect(renderSpy).to.have.been.calledWithExactly({
-        iframeUrl: '',
-        caslUrl: CASL_URL,
-        useNativeMode: true,
-      });
-    });
-
-    it('logs Showcase impression events', async () => {
-      GaaMeteringRegwall.showWithNativeRegistrationButton({
-        googleApiClientId: GOOGLE_API_CLIENT_ID,
-      });
-      await tick();
-
-      // Verify analytics events.
-      expectAnalyticsEvents([
-        {
-          analyticsEvent: AnalyticsEvent.EVENT_NO_ENTITLEMENTS,
-          isFromUserAction: false,
-        },
-        {
-          analyticsEvent: AnalyticsEvent.IMPRESSION_REGWALL,
-          isFromUserAction: false,
-        },
-        {
-          analyticsEvent: AnalyticsEvent.IMPRESSION_SHOWCASE_REGWALL,
-          isFromUserAction: false,
-        },
-      ]);
-    });
-
-    it('logs 3P button click event', async () => {
-      // Show button.
-      GaaMeteringRegwall.render_({useNativeMode: true});
-
-      GaaMeteringRegwall.createNative3PRegistrationButton({
-        authorizationUrl: GOOGLE_3P_AUTH_URL,
-      });
-
-      // Click button.
-      self.document.getElementById(GOOGLE_3P_SIGN_IN_BUTTON_ID).click();
-      clock.tick(100);
-      await tick(10);
-
-      // Verify analytics event.
-      expectAnalyticsEvents([
-        {
-          analyticsEvent:
-            AnalyticsEvent.ACTION_SHOWCASE_REGWALL_3P_BUTTON_CLICK,
-          isFromUserAction: true,
-        },
-      ]);
-    });
-  });
-
-  describe('createNativeRegistrationButton', () => {
-    it('fails if regwall is not present', async () => {
-      expect(
-        GaaMeteringRegwall.createNativeRegistrationButton({
-          googleApiClientId: GOOGLE_API_CLIENT_ID,
-        })
-      ).to.be.false;
-    });
-
-    it('renders Google Sign-In button', async () => {
-      GaaMeteringRegwall.render_({useNativeMode: true});
-      clock.tick(100);
-      await tick(10);
-
-      GaaMeteringRegwall.createNativeRegistrationButton({
-        googleApiClientId: GOOGLE_API_CLIENT_ID,
-      });
-      clock.tick(100);
-      await tick(10);
-
-      const argsInit = self.google.accounts.id.initialize.args;
-      expect(typeof argsInit[0][0].callback).to.equal('function');
-      expect(argsInit).to.deep.equal([
-        [
-          {
-            /* eslint-disable google-camelcase/google-camelcase */
-            client_id: GOOGLE_API_CLIENT_ID,
-            callback: argsInit[0][0].callback,
-            /* eslint-enable google-camelcase/google-camelcase */
-          },
-        ],
-      ]);
-
-      const argsRender = self.google.accounts.id.renderButton.args;
-      expect(argsRender).to.deep.equal([
-        [
-          self.document.getElementById(SIGN_IN_WITH_GOOGLE_BUTTON_ID),
-          {
-            'type': 'standard',
-            'theme': 'outline',
-            'text': 'continue_with',
-            'logo_alignment': 'center',
-            'click_listener': argsRender[0][1].click_listener,
-          },
-        ],
-      ]);
-    });
-
-    it('renders supported i18n languages', async () => {
-      self.document.documentElement.lang = 'pt-br';
-
-      GaaMeteringRegwall.render_({useNativeMode: true});
-      clock.tick(100);
-      await tick(10);
-
-      GaaMeteringRegwall.createNativeRegistrationButton({
-        googleApiClientId: GOOGLE_API_CLIENT_ID,
-      });
-      clock.tick(100);
-      await tick(10);
-
-      const styleEl = self.document.querySelector('style');
-      expect(styleEl.textContent).to.contain(
-        I18N_STRINGS.SHOWCASE_REGWALL_GOOGLE_SIGN_IN_BUTTON['pt-br']
-      );
-    });
-
-    it('renders "en" for non-supported i18n languages', async () => {
-      self.document.documentElement.lang = 'non-supported';
-
-      GaaMeteringRegwall.render_({useNativeMode: true});
-      clock.tick(100);
-      await tick(10);
-
-      GaaMeteringRegwall.createNativeRegistrationButton({
-        googleApiClientId: GOOGLE_API_CLIENT_ID,
-      });
-      clock.tick(100);
-      await tick(10);
-
-      const styleEl = self.document.querySelector('style');
-      expect(styleEl.textContent).to.contain(
-        I18N_STRINGS.SHOWCASE_REGWALL_GOOGLE_SIGN_IN_BUTTON['en']
-      );
-    });
-
-    it('resolves with GAA user', async () => {
-      GaaMeteringRegwall.render_({useNativeMode: true});
-      clock.tick(100);
-      await tick(10);
-
-      // Show button.
-      const gaaUserPromise = GaaMeteringRegwall.createNativeRegistrationButton({
-        googleApiClientId: GOOGLE_API_CLIENT_ID,
-      });
-      clock.tick(100);
-      await tick(10);
-
-      // Click button.
-      self.document.getElementById(SIGN_IN_WITH_GOOGLE_BUTTON_ID).click();
-
-      const args = self.google.accounts.id.initialize.args;
-      args[0][0].callback(SIGN_IN_WITH_GOOGLE_JWT);
-
-      // Send JWT.
-      expect(await gaaUserPromise).to.deep.equal(SIGN_IN_WITH_GOOGLE_JWT);
-    });
-  });
-
-  describe('createNative3PRegistrationButton', () => {
-    it('fails if regwall is not present', async () => {
-      expect(
-        GaaMeteringRegwall.createNative3PRegistrationButton({
-          authorizationUrl: GOOGLE_3P_AUTH_URL,
-        })
-      ).to.be.false;
-    });
-
-    it('renders third party Google Sign-In button', async () => {
-      GaaMeteringRegwall.render_({useNativeMode: true});
-      clock.tick(100);
-      await tick(10);
-
-      GaaMeteringRegwall.createNative3PRegistrationButton({
-        authorizationUrl: GOOGLE_3P_AUTH_URL,
-      });
-
-      const buttonDiv = self.document.querySelector(
-        '#' + GOOGLE_3P_SIGN_IN_BUTTON_ID
-      );
-      assert(buttonDiv);
-      expect(buttonDiv.tabIndex).to.equal(0);
-    });
-
-    it('renders supported i18n languages', async () => {
-      self.document.documentElement.lang = 'pt-br';
-
-      GaaMeteringRegwall.render_({useNativeMode: true});
-      clock.tick(100);
-      await tick(10);
-
-      GaaMeteringRegwall.createNative3PRegistrationButton({
-        authorizationUrl: GOOGLE_3P_AUTH_URL,
-      });
-
-      const styleEl = self.document.querySelector('style');
-      expect(styleEl.textContent).to.contain(
-        I18N_STRINGS.SHOWCASE_REGWALL_GOOGLE_SIGN_IN_BUTTON['pt-br']
-      );
-    });
-
-    it('renders English by default, if "lang" URL param is missing', async () => {
-      self.document.documentElement.lang = 'non-supported';
-
-      GaaMeteringRegwall.render_({useNativeMode: true});
-      clock.tick(100);
-      await tick(10);
-
-      GaaMeteringRegwall.createNative3PRegistrationButton({
-        authorizationUrl: GOOGLE_3P_AUTH_URL,
-      });
-
-      const styleEl = self.document.querySelector('style');
-      expect(styleEl.textContent).to.contain(
-        I18N_STRINGS.SHOWCASE_REGWALL_GOOGLE_SIGN_IN_BUTTON['en']
-      );
-    });
-
-    it('should open an authorizationUrl in the same window on click', async () => {
-      // Show button.
-      GaaMeteringRegwall.render_({useNativeMode: true});
-      clock.tick(100);
-      await tick(10);
-
-      GaaMeteringRegwall.createNative3PRegistrationButton({
-        authorizationUrl: GOOGLE_3P_AUTH_URL,
-      });
-
-      // Click button.
-      self.document.getElementById(GOOGLE_3P_SIGN_IN_BUTTON_ID).click();
-      clock.tick(100);
-      await tick(10);
-
-      expect(self.open).to.have.been.calledWithExactly(
-        GOOGLE_3P_AUTH_URL,
-        '_parent'
-      );
-    });
-  });
-
-  describe('remove', () => {
-    it('removes regwall', () => {
-      function findRegwallInDocument() {
-        return self.document.querySelector('.gaa-metering-regwall--dialog');
-      }
-
-      GaaMeteringRegwall.show({iframeUrl: GSI_IFRAME_URL});
-      expect(findRegwallInDocument()).to.not.be.null;
-
-      GaaMeteringRegwall.remove();
-      expect(findRegwallInDocument()).to.be.null;
-    });
-  });
-
-  describe('signOut', () => {
-    it('tells GSI to sign user out', async () => {
-      const promise = GaaMeteringRegwall.signOut();
-      clock.tick(100);
-      await promise;
-
-      expect(signOutFake).to.be.called;
-    });
-  });
-
-  describe('getGaaUser_', () => {
-    it('sends intro postMessage to iframe', async () => {
-      // Mock an iframe (as a div to avoid browser security restrictions).
-      const mockIframeEl = self.document.createElement('div');
-      mockIframeEl.id = GOOGLE_SIGN_IN_IFRAME_ID;
-      mockIframeEl.contentWindow = {
-        postMessage: sandbox.fake(),
-      };
-      self.document.body.appendChild(mockIframeEl);
-
-      // Send intro message then trigger mock iframe's onload callback.
-      GaaMeteringRegwall.sendIntroMessageToGsiIframe_({
-        iframeUrl: GSI_IFRAME_URL,
-      });
-      mockIframeEl.onload();
-
-      expect(mockIframeEl.contentWindow.postMessage).to.be.calledWithExactly(
-        {
-          command: POST_MESSAGE_COMMAND_INTRODUCTION,
-          stamp: POST_MESSAGE_STAMP,
-        },
-        'https://localhost'
-      );
-    });
-  });
-
-  describe('logButtonClickEvents_', () => {
-    it('sends GSI button click event', async () => {
-      // Show Regwall.
-      GaaMeteringRegwall.show({iframeUrl: GSI_IFRAME_URL});
-      await tick();
-      logEvent.resetHistory();
-
-      // Send button click post message.
-      postMessage({
-        stamp: POST_MESSAGE_STAMP,
-        command: POST_MESSAGE_COMMAND_GSI_BUTTON_CLICK,
-      });
-
-      // Wait for logging.
-      await new Promise((resolve) => {
-        logEvent = sandbox.fake(resolve);
-      });
-
-      // Verify analytics event.
-      expectAnalyticsEvents([
-        {
-          analyticsEvent: AnalyticsEvent.ACTION_SHOWCASE_REGWALL_GSI_CLICK,
-          isFromUserAction: true,
-        },
-      ]);
-    });
-
-    it('sends SIWG button click event', async () => {
-      // Show Regwall.
-      GaaMeteringRegwall.show({iframeUrl: SIWG_IFRAME_URL});
-      await tick();
-      logEvent.resetHistory();
-
-      // Send button click post message.
-      postMessage({
-        stamp: POST_MESSAGE_STAMP,
-        command: POST_MESSAGE_COMMAND_SIWG_BUTTON_CLICK,
-      });
-
-      // Wait for logging.
-      await new Promise((resolve) => {
-        logEvent = sandbox.fake(resolve);
-      });
-
-      // Verify analytics event.
-      expectAnalyticsEvents([
-        {
-          analyticsEvent: AnalyticsEvent.ACTION_SHOWCASE_REGWALL_SIWG_CLICK,
-          isFromUserAction: true,
-        },
-      ]);
-    });
-
-    it('sends 3P button click event', async () => {
-      // Show Regwall.
-      GaaMeteringRegwall.show({iframeUrl: GSI_IFRAME_URL});
-      await tick();
-      logEvent.resetHistory();
-
-      // Send button click post message.
-      postMessage({
-        stamp: POST_MESSAGE_STAMP,
-        command: POST_MESSAGE_COMMAND_3P_BUTTON_CLICK,
-      });
-
-      // Wait for logging.
-      await new Promise((resolve) => {
-        logEvent = sandbox.fake(resolve);
-      });
-
-      // Verify analytics event.
-      expectAnalyticsEvents([
-        {
-          analyticsEvent:
-            AnalyticsEvent.ACTION_SHOWCASE_REGWALL_3P_BUTTON_CLICK,
-          isFromUserAction: true,
-        },
-      ]);
-    });
-  });
-
-  describe('getPublisherNameFromPageConfig_', () => {
-    it('gets the publisher name from object page config', () => {
-      expect(GaaMeteringRegwall.getPublisherNameFromPageConfig_()).to.equal(
-        PUBLISHER_NAME
-      );
-    });
-
-    it('gets the publisher name from array page config', () => {
-      self.document.head.innerHTML = `
-        <script type="application/ld+json">
-          [${ARTICLE_LD_JSON_METADATA}]
-        </script>
-      `;
-
-      expect(GaaMeteringRegwall.getPublisherNameFromPageConfig_()).to.equal(
-        PUBLISHER_NAME
-      );
-    });
-
-    it('gets the publisher name from graph construct', () => {
-      self.document.head.innerHTML = `
-        <script type="application/ld+json">
-          [{
-            "@context": "http://schema.org",
-            "@graph": [${ARTICLE_LD_JSON_METADATA}]
-          }]
-        </script>
-      `;
-
-      expect(GaaMeteringRegwall.getPublisherNameFromPageConfig_()).to.equal(
-        PUBLISHER_NAME
-      );
-    });
-  });
-});
-
-describes.realWin('GaaGoogleSignInButton', () => {
-  const allowedOrigins = [location.origin];
-
-  let clock;
-
-  beforeEach(() => {
-    // Mock clock.
-    clock = sandbox.useFakeTimers();
-
-    self.gapi = {
-      load: sandbox.fake((name, callback) => {
-        callback();
-      }),
-      auth2: {
-        init: sandbox.fake(),
-        getAuthInstance: sandbox.fake(),
-      },
-      signin2: {
-        render: sandbox.fake(),
-      },
-    };
-
-    // Mock query string.
-    sandbox.stub(QueryStringUtils, 'getQueryString');
-    QueryStringUtils.getQueryString.returns('?lang=en');
-
-    // Mock console.warn method.
-    sandbox.stub(self.console, 'warn');
-  });
-
-  afterEach(() => {
-    if (self.postMessage.restore) {
-      self.postMessage.restore();
-    }
-    QueryStringUtils.getQueryString.restore();
-
-    // Remove the injected style from GaaGoogleSignInButton.show.
-    for (const style of [...self.document.head.querySelectorAll('style')]) {
-      style.remove();
-    }
-
-    self.console.warn.restore();
-  });
-
-  describe('show', () => {
-    it('renders Google Sign-In button', async () => {
-      GaaGoogleSignInButton.show({allowedOrigins});
-      clock.tick(100);
-      await tick(10);
-
-      const args = self.gapi.signin2.render.args;
-      expect(typeof args[0][1].onsuccess).to.equal('function');
-      expect(args).to.deep.equal([
-        [
-          'swg-google-sign-in-button',
-          {
-            longtitle: true,
-            onsuccess: args[0][1].onsuccess,
-            prompt: 'select_account',
-            scope: 'profile email',
-            theme: 'dark',
-          },
-        ],
-      ]);
-    });
-
-    it('renders supported i18n languages', async () => {
-      QueryStringUtils.getQueryString.returns('?lang=pt-br');
-
-      GaaGoogleSignInButton.show({allowedOrigins});
-      clock.tick(100);
-      await tick(10);
-
-      const styleEl = self.document.querySelector('style');
-      expect(styleEl.textContent).to.contain(
-        I18N_STRINGS.SHOWCASE_REGWALL_GOOGLE_SIGN_IN_BUTTON['pt-br']
-      );
-    });
-
-    it('renders English by default, if "lang" URL param is missing', async () => {
-      QueryStringUtils.getQueryString.returns('?');
-
-      GaaGoogleSignInButton.show({allowedOrigins});
-      clock.tick(100);
-      await tick(10);
-
-      const styleEl = self.document.querySelector('style');
-      expect(styleEl.textContent).to.contain(
-        I18N_STRINGS.SHOWCASE_REGWALL_GOOGLE_SIGN_IN_BUTTON['en']
-      );
-    });
-
-    it('sends post message with button click event', async () => {
-      // Show button.
-      GaaGoogleSignInButton.show({allowedOrigins});
-
-      // Send intro post message.
-      postMessage({
-        stamp: POST_MESSAGE_STAMP,
-        command: POST_MESSAGE_COMMAND_INTRODUCTION,
-      });
-
-      // Wait for promises and intervals to resolve.
-      clock.tick(100);
-      await tick(10);
-
-      // Click button.
-      self.document.getElementById(GOOGLE_SIGN_IN_BUTTON_ID).click();
-
-      // Wait for button click post message.
-      await new Promise((resolve) => {
-        sandbox.stub(self, 'postMessage').callsFake(() => {
-          resolve();
-        });
-      });
-
-      // Expect button click post message.
-      expect(self.postMessage).to.be.calledWithExactly(
-        {
-          command: POST_MESSAGE_COMMAND_GSI_BUTTON_CLICK,
-          stamp: POST_MESSAGE_STAMP,
-        },
-        location.origin
-      );
-    });
-
-    it('sends post message with GAA user', async () => {
-      GaaGoogleSignInButton.show({allowedOrigins});
-
-      // Send intro post message.
-      postMessage({
-        stamp: POST_MESSAGE_STAMP,
-        command: POST_MESSAGE_COMMAND_INTRODUCTION,
-      });
-
-      // Mock Google Sign-In response with GoogleUser object.
-      const gaaUser = {
-        idToken: 'idToken',
-        name: 'name',
-        givenName: 'givenName',
-        familyName: 'familyName',
-        imageUrl: 'imageUrl',
-        email: 'email',
-        authorizationData: {
-          /* eslint-disable google-camelcase/google-camelcase */
-          access_token: 'accessToken',
-          id_token: 'idToken',
-          scope: 'scope',
-          expires_in: 0,
-          first_issued_at: 0,
-          expires_at: 0,
-          /* eslint-enable google-camelcase/google-camelcase */
-        },
-      };
-      const googleUser = {
-        getBasicProfile: () => ({
-          getName: () => gaaUser.name,
-          getGivenName: () => gaaUser.givenName,
-          getFamilyName: () => gaaUser.familyName,
-          getImageUrl: () => gaaUser.imageUrl,
-          getEmail: () => gaaUser.email,
-        }),
-        getAuthResponse: () => gaaUser.authorizationData,
-      };
-      const args = self.gapi.signin2.render.args;
-      // Wait for promises and intervals to resolve.
-      clock.tick(100);
-      await tick(10);
-      // Send GoogleUser.
-      args[0][1].onsuccess(googleUser);
-
-      // Wait for post message.
-      await new Promise((resolve) => {
-        sandbox.stub(self, 'postMessage').callsFake(() => {
-          resolve();
-        });
-      });
-
-      expect(self.postMessage).to.be.calledWithExactly(
-        {
-          command: POST_MESSAGE_COMMAND_USER,
-          gaaUser: {
-            email: 'email',
-            familyName: 'familyName',
-            givenName: 'givenName',
-            idToken: 'idToken',
-            imageUrl: 'imageUrl',
-            name: 'name',
-            authorizationData: {
-              /* eslint-disable google-camelcase/google-camelcase */
-              access_token: 'accessToken',
-              id_token: 'idToken',
-              scope: 'scope',
-              expires_in: 0,
-              first_issued_at: 0,
-              expires_at: 0,
-              /* eslint-enable google-camelcase/google-camelcase */
-            },
-          },
-          stamp: POST_MESSAGE_STAMP,
-        },
-        location.origin
-      );
-    });
-
-    it('sends errors to parent', async () => {
-      self.gapi.signin2.render = sandbox.fake.throws('I need cookies');
-
-      GaaGoogleSignInButton.show({allowedOrigins});
-
-      // Send intro post message.
-      postMessage({
-        stamp: POST_MESSAGE_STAMP,
-        command: POST_MESSAGE_COMMAND_INTRODUCTION,
-      });
-
-      // Wait for promises and intervals to resolve.
-      clock.tick(100);
-      await tick(10);
-
-      // Wait for post message.
-      await new Promise((resolve) => {
-        sandbox.stub(self, 'postMessage').callsFake(() => {
-          resolve();
-        });
-      });
-
-      expect(self.postMessage).to.be.calledWithExactly(
-        {
-          command: POST_MESSAGE_COMMAND_ERROR,
-          stamp: POST_MESSAGE_STAMP,
-        },
-        location.origin
-      );
-    });
-
-    it('fails and warns when passed invalid origins', async () => {
-      const invalidOrigins = [
-        // Bad protocol, should be http or https.
-        'ftp://localhost:8080',
-        // Includes path.
-        'http://localhost:8080/',
-      ];
-
-      for (const invalidOrigin of invalidOrigins) {
-        GaaGoogleSignInButton.show({allowedOrigins: [invalidOrigin]});
-
-        // Send intro post message.
-        postMessage({
-          stamp: POST_MESSAGE_STAMP,
-          command: POST_MESSAGE_COMMAND_INTRODUCTION,
-        });
-
-        // Wait for promises and intervals to resolve.
-        clock.tick(100);
-        await tick(10);
-
-        expect(self.console.warn).to.have.been.calledWithExactly(
-          `[swg-gaa.js:GaaGoogleSignInButton.show]: You specified an invalid origin: ${invalidOrigin}`
-        );
-      }
-    });
-  });
-});
-
-describes.realWin('GaaSignInWithGoogleButton', () => {
-  const allowedOrigins = [location.origin];
-  const clientId = 'client_id';
-
-  let clock;
-
-  beforeEach(() => {
-    // Mock clock.
-    clock = sandbox.useFakeTimers();
-
-    self.google = {
-      accounts: {
-        id: {
-          initialize: sandbox.fake(),
-          renderButton: sandbox.fake(),
-        },
-      },
-    };
-
-    // Mock query string.
-    sandbox.stub(QueryStringUtils, 'getQueryString');
-    QueryStringUtils.getQueryString.returns('?lang=en');
-
-    // Mock console.warn method.
-    sandbox.stub(self.console, 'warn');
-  });
-
-  afterEach(() => {
-    if (self.postMessage.restore) {
-      self.postMessage.restore();
-    }
-    QueryStringUtils.getQueryString.restore();
-
-    // Remove the injected style from GaaSignInWithGoogleButton.show.
-    for (const style of [...self.document.head.querySelectorAll('style')]) {
-      style.remove();
-    }
-
-    self.console.warn.restore();
-  });
-
-  describe('show', () => {
-    it('renders Sign-In with Google button', async () => {
-      GaaSignInWithGoogleButton.show({clientId, allowedOrigins});
-      clock.tick(100);
-      await tick(10);
-
-      const argsInit = self.google.accounts.id.initialize.args;
-      expect(typeof argsInit[0][0].callback).to.equal('function');
-      expect(argsInit).to.deep.equal([
-        [
-          {
-            /* eslint-disable google-camelcase/google-camelcase */
-            client_id: clientId,
-            callback: argsInit[0][0].callback,
-            allowed_parent_origin: allowedOrigins,
-            /* eslint-enable google-camelcase/google-camelcase */
-          },
-        ],
-      ]);
-
-      const argsRender = self.google.accounts.id.renderButton.args;
-      const buttonEl = self.document.getElementById(
-        SIGN_IN_WITH_GOOGLE_BUTTON_ID
-      );
-
-      expect(argsRender).to.deep.equal([
-        [
-          buttonEl,
-          {
-            'type': 'standard',
-            'theme': 'outline',
-            'text': 'continue_with',
-            'logo_alignment': 'center',
-            'width': buttonEl.offsetWidth,
-            'height': buttonEl.offsetHeight,
-            'click_listener': argsRender[0][1].click_listener,
-          },
-        ],
-      ]);
-    });
-
-    it('renders supported i18n languages', async () => {
-      QueryStringUtils.getQueryString.returns('?lang=pt-br');
-
-      GaaSignInWithGoogleButton.show({clientId, allowedOrigins});
-      clock.tick(100);
-      await tick(10);
-
-      const styleEl = self.document.querySelector('style');
-      expect(styleEl.textContent).to.contain(
-        I18N_STRINGS.SHOWCASE_REGWALL_GOOGLE_SIGN_IN_BUTTON['pt-br']
-      );
-    });
-
-    it('renders English by default, if "lang" URL param is missing', async () => {
-      QueryStringUtils.getQueryString.returns('?');
-
-      GaaSignInWithGoogleButton.show({clientId, allowedOrigins});
-      clock.tick(100);
-      await tick(10);
-
-      const styleEl = self.document.querySelector('style');
-      expect(styleEl.textContent).to.contain(
-        I18N_STRINGS.SHOWCASE_REGWALL_GOOGLE_SIGN_IN_BUTTON['en']
-      );
-    });
-
-    it('sends post message with button click event', async () => {
-      // Show button.
-      GaaSignInWithGoogleButton.show({clientId, allowedOrigins});
-
-      // Send intro post message.
-      postMessage({
-        stamp: POST_MESSAGE_STAMP,
-        command: POST_MESSAGE_COMMAND_INTRODUCTION,
-      });
-
-      // Wait for promises and intervals to resolve.
-      clock.tick(100);
-      await tick(10);
-
-      // Simulate a click event from SIWG.
-      self.google.accounts.id.renderButton.args[0][1].click_listener();
-
-      await new Promise((resolve) => {
-        sandbox.stub(self, 'postMessage').callsFake(() => {
-          resolve();
-        });
-      });
-
-      // Expect button click post message.
-      expect(self.postMessage).to.be.calledWithExactly(
-        {
-          command: POST_MESSAGE_COMMAND_SIWG_BUTTON_CLICK,
-          stamp: POST_MESSAGE_STAMP,
-        },
-        location.origin
-      );
-    });
-
-    it('sends post message with GAA user', async () => {
-      // Mock encrypted GIS response with JWT object.
-      const jwtRaw = {
-        credential: {
-          payload:
-            'eyJpc3MiOiJodHRwczovL2FjY291bnRzLmdvb2dsZS5jb20iLCJuYmYiOjE2NDE5OTU5MzgsImF1ZCI6IjQ3MzExNjQ0Mzk1OC12ajkwaDJrbW92cm9zZXUydWhrdnVxNGNjZ3ZldW43My5hcHBzLmdvb2dsZXVzZXJjb250ZW50LmNvbSIsInN1YiI6IjExNzE5ODI3Njk2NjYyOTcxNjg0MSIsImhkIjoiZ29vZ2xlLmNvbSIsImVtYWlsIjoiZWRiaXJkQGdvb2dsZS5jb20iLCJlbWFpbF92ZXJpZmllZCI6dHJ1ZSwiYXpwIjoiNDczMTE2NDQzOTU4LXZqOTBoMmttb3Zyb3NldTJ1aGt2dXE0Y2NndmV1bjczLmFwcHMuZ29vZ2xldXNlcmNvbnRlbnQuY29tIiwiaWF0IjoxNjQxOTk2MjM4LCJleHAiOjE2NDE5OTk4MzgsImp0aSI6IjA4MGQ3Y2FiNzAyZDEyYWU1MzJjYzc3YTExNDk3NGI4OThjNmFjNTYifQ',
-        },
-      };
-
-      // Mock decrypted GIS response with JWT object.
-      const jwtDecoded = {
-        credential: {
-          /* eslint-disable google-camelcase/google-camelcase */
-          payload: {
-            iss: 'https://accounts.google.com', // The JWT's issuer
-            nbf: 161803398874,
-            aud: '314159265-pi.apps.googleusercontent.com', // Your server's client ID
-            sub: '3141592653589793238', // The unique ID of the user's Google Account
-            hd: 'gmail.com', // If present, the host domain of the user's GSuite email address
-            email: 'elisa.g.beckett@gmail.com', // The user's email address
-            email_verified: true, // true, if Google has verified the email address
-            azp: '314159265-pi.apps.googleusercontent.com',
-            name: 'Elisa Beckett',
-            // If present, a URL to user's profile picture
-            picture:
-              'https://lh3.googleusercontent.com/a-/e2718281828459045235360uler',
-            given_name: 'Elisa',
-            family_name: 'Beckett',
-            iat: 1596474000, // Unix timestamp of the assertion's creation time
-            exp: 1596477600, // Unix timestamp of the assertion's expiration time
-            jti: 'abc161803398874def',
-          },
-          /* eslint-enable google-camelcase/google-camelcase */
-        },
-      };
-
-      // Mock JWT decoding function.
-      sandbox.stub(JwtHelper.prototype, 'decode').callsFake((unused) => {
-        return jwtDecoded.credential;
-      });
-
-      GaaSignInWithGoogleButton.show({
-        clientId,
-        allowedOrigins,
-      });
-
-      // Send intro post message.
-      postMessage({
-        stamp: POST_MESSAGE_STAMP,
-        command: POST_MESSAGE_COMMAND_INTRODUCTION,
-      });
-
-      const args = self.google.accounts.id.initialize.args;
-      // Wait for promises and intervals to resolve.
-      clock.tick(100);
-      await tick(10);
-      // Send JWT.
-      args[0][0].callback(jwtRaw);
-
-      // Wait for post message.
-      await new Promise((resolve) => {
-        sandbox.stub(self, 'postMessage').callsFake(() => {
-          resolve();
-        });
-      });
-
-      expect(self.postMessage).to.be.calledWithExactly(
-        {
-          command: POST_MESSAGE_COMMAND_USER,
-          jwtPayload: jwtDecoded.credential,
-          returnedJwt: jwtDecoded.credential,
-          stamp: POST_MESSAGE_STAMP,
-        },
-        location.origin
-      );
-    });
-
-    it('sends post message with GAA user while requesting raw JWT', async () => {
-      // Mock encrypted GIS response with JWT object.
-      const jwtRaw = {
-        credential: {
-          payload:
-            'eyJpc3MiOiJodHRwczovL2FjY291bnRzLmdvb2dsZS5jb20iLCJuYmYiOjE2NDE5OTU5MzgsImF1ZCI6IjQ3MzExNjQ0Mzk1OC12ajkwaDJrbW92cm9zZXUydWhrdnVxNGNjZ3ZldW43My5hcHBzLmdvb2dsZXVzZXJjb250ZW50LmNvbSIsInN1YiI6IjExNzE5ODI3Njk2NjYyOTcxNjg0MSIsImhkIjoiZ29vZ2xlLmNvbSIsImVtYWlsIjoiZWRiaXJkQGdvb2dsZS5jb20iLCJlbWFpbF92ZXJpZmllZCI6dHJ1ZSwiYXpwIjoiNDczMTE2NDQzOTU4LXZqOTBoMmttb3Zyb3NldTJ1aGt2dXE0Y2NndmV1bjczLmFwcHMuZ29vZ2xldXNlcmNvbnRlbnQuY29tIiwiaWF0IjoxNjQxOTk2MjM4LCJleHAiOjE2NDE5OTk4MzgsImp0aSI6IjA4MGQ3Y2FiNzAyZDEyYWU1MzJjYzc3YTExNDk3NGI4OThjNmFjNTYifQ',
-        },
-      };
-
-      // Mock decrypted GIS response with JWT object.
-      const jwtDecoded = {
-        credential: {
-          /* eslint-disable google-camelcase/google-camelcase */
-          payload: {
-            iss: 'https://accounts.google.com', // The JWT's issuer
-            nbf: 161803398874,
-            aud: '314159265-pi.apps.googleusercontent.com', // Your server's client ID
-            sub: '3141592653589793238', // The unique ID of the user's Google Account
-            hd: 'gmail.com', // If present, the host domain of the user's GSuite email address
-            email: 'elisa.g.beckett@gmail.com', // The user's email address
-            email_verified: true, // true, if Google has verified the email address
-            azp: '314159265-pi.apps.googleusercontent.com',
-            name: 'Elisa Beckett',
-            // If present, a URL to user's profile picture
-            picture:
-              'https://lh3.googleusercontent.com/a-/e2718281828459045235360uler',
-            given_name: 'Elisa',
-            family_name: 'Beckett',
-            iat: 1596474000, // Unix timestamp of the assertion's creation time
-            exp: 1596477600, // Unix timestamp of the assertion's expiration time
-            jti: 'abc161803398874def',
-          },
-          /* eslint-enable google-camelcase/google-camelcase */
-        },
-      };
-
-      // Mock JWT decoding function.
-      sandbox.stub(JwtHelper.prototype, 'decode').callsFake((unused) => {
-        return jwtDecoded.credential;
-      });
-
-      GaaSignInWithGoogleButton.show({
-        clientId,
-        allowedOrigins,
-        rawJwt: true,
-      });
-
-      // Send intro post message.
-      postMessage({
-        stamp: POST_MESSAGE_STAMP,
-        command: POST_MESSAGE_COMMAND_INTRODUCTION,
-      });
-
-      const args = self.google.accounts.id.initialize.args;
-      // Wait for promises and intervals to resolve.
-      clock.tick(100);
-      await tick(10);
-      // Send JWT.
-      args[0][0].callback(jwtRaw);
-
-      // Wait for post message.
-      await new Promise((resolve) => {
-        sandbox.stub(self, 'postMessage').callsFake(() => {
-          resolve();
-        });
-      });
-
-      expect(self.postMessage).to.be.calledWithExactly(
-        {
-          command: POST_MESSAGE_COMMAND_USER,
-          jwtPayload: jwtDecoded.credential,
-          returnedJwt: jwtRaw,
-          stamp: POST_MESSAGE_STAMP,
-        },
-        location.origin
-      );
-    });
-
-    it('sends errors to parent', async () => {
-      self.google.accounts.id.initialize = sandbox.fake.throws(
-        'Function not loaded'
-      );
-
-      GaaSignInWithGoogleButton.show({clientId, allowedOrigins});
-
-      // Send intro post message.
-      postMessage({
-        stamp: POST_MESSAGE_STAMP,
-        command: POST_MESSAGE_COMMAND_INTRODUCTION,
-      });
-
-      // Wait for promises and intervals to resolve.
-      clock.tick(100);
-      await tick(10);
-
-      // Wait for post message.
-      await new Promise((resolve) => {
-        sandbox.stub(self, 'postMessage').callsFake(() => {
-          resolve();
-        });
-      });
-
-      expect(self.postMessage).to.be.calledWithExactly(
-        {
-          command: POST_MESSAGE_COMMAND_ERROR,
-          stamp: POST_MESSAGE_STAMP,
-        },
-        location.origin
-      );
-    });
-
-    it('fails and warns when passed invalid origins', async () => {
-      const invalidOrigins = [
-        // Bad protocol, should be http or https.
-        'ftp://localhost:8080',
-        // Includes path.
-        'http://localhost:8080/',
-      ];
-
-      for (const invalidOrigin of invalidOrigins) {
-        GaaSignInWithGoogleButton.show({
-          clientId,
-          allowedOrigins: [invalidOrigin],
-        });
-
-        // Send intro post message.
-        postMessage({
-          stamp: POST_MESSAGE_STAMP,
-          command: POST_MESSAGE_COMMAND_INTRODUCTION,
-        });
-
-        // Wait for promises and intervals to resolve.
-        clock.tick(100);
-        await tick(10);
-
-        expect(self.console.warn).to.have.been.calledWithExactly(
-          `[swg-gaa.js:GaaSignInWithGoogleButton.show]: You specified an invalid origin: ${invalidOrigin}`
-        );
-      }
-    });
-  });
-});
-
-describes.realWin('GaaGoogle3pSignInButton', () => {
-  const allowedOrigins = [location.origin];
-
-  let clock;
-
-  beforeEach(() => {
-    // Mock clock.
-    clock = sandbox.useFakeTimers();
-
-    // Mock query string.
-    sandbox.stub(QueryStringUtils, 'getQueryString');
-    QueryStringUtils.getQueryString.returns('?lang=en');
-
-    // Mock console.warn method.
-    sandbox.stub(self.console, 'warn');
-
-    sandbox.stub(self, 'open');
-    // Makes sure no div elements on the test dom exist.
-    const elements = self.document.getElementsByTagName('div');
-    while (elements[0]) {
-      elements[0].parentNode.removeChild(elements[0]);
-    }
-  });
-
-  afterEach(() => {
-    if (self.postMessage.restore) {
-      self.postMessage.restore();
-    }
-
-    QueryStringUtils.getQueryString.restore();
-
-    // Remove the injected style from GaaGoogle3pSignInButton.show.
-    for (const style of [...self.document.head.querySelectorAll('style')]) {
-      style.remove();
-    }
-
-    self.document.getElementById(GOOGLE_3P_SIGN_IN_BUTTON_ID)?.remove();
-    self.console.warn.restore();
-  });
-
-  describe('show', () => {
-    it('sends errors to parent', async () => {
-      const invalidOrigin = [
-        // Bad protocol, should be http or https.
-        'ftp://localhost:8080',
-        location.origin,
-      ];
-
-      GaaGoogle3pSignInButton.show(
-        {allowedOrigins: invalidOrigin},
-        GOOGLE_3P_AUTH_URL
-      );
-
-      // Send intro post message.
-      postMessage({
-        stamp: POST_MESSAGE_STAMP,
-        command: POST_MESSAGE_COMMAND_INTRODUCTION,
-      });
-
-      // Wait for promises and intervals to resolve.
-      clock.tick(100);
-      await tick(10);
-
-      // Wait for post message.
-      await new Promise((resolve) => {
-        sandbox.stub(self, 'postMessage').callsFake(() => {
-          resolve();
-        });
-      });
-
-      expect(self.postMessage).to.be.calledWith(
-        {
-          command: POST_MESSAGE_COMMAND_ERROR,
-          stamp: POST_MESSAGE_STAMP,
-        },
-        location.origin
-      );
-    });
-
-    it('fails and warns when passed invalid origins', async () => {
-      const invalidOrigins = [
-        // Bad protocol, should be http or https.
-        'ftp://localhost:8080',
-        // Includes path.
-        'http://localhost:8080/',
-      ];
-
-      for (const invalidOrigin of invalidOrigins) {
-        GaaGoogle3pSignInButton.show(
-          {allowedOrigins: [invalidOrigin]},
-          GOOGLE_3P_AUTH_URL
-        );
-
-        // Send intro post message.
-        postMessage({
-          stamp: POST_MESSAGE_STAMP,
-          command: POST_MESSAGE_COMMAND_INTRODUCTION,
-        });
-
-        // Wait for promises and intervals to resolve.
-        clock.tick(100);
-        await tick(10);
-
-        expect(self.console.warn).to.have.been.calledWithExactly(
-          `[swg-gaa.js:GaaGoogle3pSignInButton.show]: You specified an invalid origin: ${invalidOrigin}`
-        );
-      }
-    });
-
-    it('renders third party Google Sign-In button', async () => {
-      GaaGoogle3pSignInButton.show({allowedOrigins}, GOOGLE_3P_AUTH_URL);
-      clock.tick(100);
-      await tick(10);
-
-      const buttonDiv = self.document.querySelector(
-        '#' + GOOGLE_3P_SIGN_IN_BUTTON_ID
-      );
-      assert(buttonDiv);
-      expect(buttonDiv.tabIndex).to.equal(0);
-      expect(typeof buttonDiv.onclick).to.equal('function');
-    });
-
-    it('renders supported i18n languages', async () => {
-      QueryStringUtils.getQueryString.returns('?lang=pt-br');
-
-      GaaGoogle3pSignInButton.show({allowedOrigins}, GOOGLE_3P_AUTH_URL);
-      clock.tick(100);
-      await tick(10);
-
-      const styleEl = self.document.querySelector('style');
-      expect(styleEl.textContent).to.contain(
-        I18N_STRINGS.SHOWCASE_REGWALL_GOOGLE_SIGN_IN_BUTTON['pt-br']
-      );
-    });
-
-    it('renders English by default, if "lang" URL param is missing', async () => {
-      QueryStringUtils.getQueryString.returns('?');
-
-      GaaGoogle3pSignInButton.show({allowedOrigins}, GOOGLE_3P_AUTH_URL);
-      clock.tick(100);
-      await tick(10);
-
-      const styleEl = self.document.querySelector('style');
-      expect(styleEl.textContent).to.contain(
-        I18N_STRINGS.SHOWCASE_REGWALL_GOOGLE_SIGN_IN_BUTTON['en']
-      );
-    });
-
-    it('sends post message with button click event', async () => {
-      // Show button.
-      GaaGoogle3pSignInButton.show({
-        allowedOrigins,
-        authorizationUrl: GOOGLE_3P_AUTH_URL,
-      });
-      clock.tick(100);
-      await tick(10);
-
-      // Send intro post message.
-      postMessage({
-        stamp: POST_MESSAGE_STAMP,
-        command: POST_MESSAGE_COMMAND_INTRODUCTION,
-      });
-
-      // Click button.
-      self.document.getElementById(GOOGLE_3P_SIGN_IN_BUTTON_ID).click();
-      clock.tick(100);
-      await tick(10);
-
-      // Send user post message.
-      postMessage({
-        stamp: POST_MESSAGE_STAMP,
-        command: POST_MESSAGE_COMMAND_USER,
-      });
-
-      // Wait for promises and intervals to resolve.
-      clock.tick(100);
-      await tick(10);
-
-      // Wait for `postMessage` to be relayed to parent.
-      await new Promise((resolve) => {
-        sandbox.stub(self.parent, 'postMessage').callsFake(() => {
-          resolve();
-        });
-      });
-
-      expect(self.parent.postMessage).to.be.calledWithExactly(
-        {
-          stamp: POST_MESSAGE_STAMP,
-          command: POST_MESSAGE_COMMAND_USER,
-        },
-        location.origin
-      );
-    });
-
-    it('sends post message with 3p button click event', async () => {
-      // Show button.
-      GaaGoogle3pSignInButton.show({
-        allowedOrigins,
-        authorizationUrl: GOOGLE_3P_AUTH_URL,
-      });
-      clock.tick(100);
-      await tick(10);
-
-      // Send intro post message.
-      postMessage({
-        stamp: POST_MESSAGE_STAMP,
-        command: POST_MESSAGE_COMMAND_INTRODUCTION,
-      });
-
-      // Wait for promises and intervals to resolve.
-      clock.tick(100);
-      await tick(10);
-
-      // Click button.
-      self.document.getElementById(GOOGLE_3P_SIGN_IN_BUTTON_ID).click();
-
-      // Wait for button click post message.
-      await new Promise((resolve) => {
-        sandbox.stub(self, 'postMessage').callsFake(() => {
-          resolve();
-        });
-      });
-
-      // Expect button click post message.
-      expect(self.postMessage).to.be.calledWithExactly(
-        {
-          command: POST_MESSAGE_COMMAND_3P_BUTTON_CLICK,
-          stamp: POST_MESSAGE_STAMP,
-        },
-        location.origin
-      );
-    });
-
-    it('sends post message with 3p button click event when redirectMode is true', async () => {
-      // Show button.
-      GaaGoogle3pSignInButton.show({
-        allowedOrigins,
-        authorizationUrl: GOOGLE_3P_AUTH_URL,
-        redirectMode: true,
-      });
-      clock.tick(100);
-      await tick(10);
-
-      // Send intro post message.
-      postMessage({
-        stamp: POST_MESSAGE_STAMP,
-        command: POST_MESSAGE_COMMAND_INTRODUCTION,
-      });
-
-      // Wait for promises and intervals to resolve.
-      clock.tick(100);
-      await tick(10);
-
-      // Click button.
-      self.document.getElementById(GOOGLE_3P_SIGN_IN_BUTTON_ID).click();
-
-      // Wait for button click post message.
-      await new Promise((resolve) => {
-        sandbox.stub(self, 'postMessage').callsFake(() => {
-          resolve();
-        });
-      });
-
-      // Expect button click post message.
-      expect(self.postMessage).to.be.calledWithExactly(
-        {
-          command: POST_MESSAGE_COMMAND_3P_BUTTON_CLICK,
-          stamp: POST_MESSAGE_STAMP,
-        },
-        location.origin
-      );
-    });
-
-    it('should open an authorizationUrl in a new window by default', async () => {
-      // Show button.
-      GaaGoogle3pSignInButton.show({
-        allowedOrigins,
-        authorizationUrl: GOOGLE_3P_AUTH_URL,
-      });
-      clock.tick(100);
-      await tick(10);
-
-      // Send intro post message.
-      postMessage({
-        stamp: POST_MESSAGE_STAMP,
-        command: POST_MESSAGE_COMMAND_INTRODUCTION,
-      });
-
-      // Click button.
-      self.document.getElementById(GOOGLE_3P_SIGN_IN_BUTTON_ID).click();
-      clock.tick(100);
-      await tick(10);
-
-      // Wait for `open` to be called.
-      await new Promise((resolve) => void self.open.callsFake(resolve));
-
-      expect(self.open).to.have.been.calledWithExactly(GOOGLE_3P_AUTH_URL);
-    });
-
-    it('should open an authorizationUrl in the same window when redirectMode is true', async () => {
-      // Show button.
-      GaaGoogle3pSignInButton.show({
-        allowedOrigins,
-        authorizationUrl: GOOGLE_3P_AUTH_URL,
-        redirectMode: true,
-      });
-      clock.tick(100);
-      await tick(10);
-
-      // Send intro post message.
-      postMessage({
-        stamp: POST_MESSAGE_STAMP,
-        command: POST_MESSAGE_COMMAND_INTRODUCTION,
-      });
-
-      // Click button.
-      self.document.getElementById(GOOGLE_3P_SIGN_IN_BUTTON_ID).click();
-      clock.tick(100);
-      await tick(100);
-
-      // Wait for `open` to be called.
-      await new Promise((resolve) => void self.open.callsFake(resolve));
-
-      expect(self.open).to.have.been.calledWithExactly(
-        GOOGLE_3P_AUTH_URL,
-        '_parent'
-      );
-    });
-
-    it('should open window twice when redirectMode is true', async () => {
-      // Show button.
-      GaaGoogle3pSignInButton.show({
-        allowedOrigins,
-        authorizationUrl: GOOGLE_3P_AUTH_URL,
-        redirectMode: true,
-      });
-
-      // Send intro post message.
-      postMessage({
-        stamp: POST_MESSAGE_STAMP,
-        command: POST_MESSAGE_COMMAND_INTRODUCTION,
-      });
-
-      // Click button.
-      self.document.getElementById(GOOGLE_3P_SIGN_IN_BUTTON_ID).click();
-
-      // Wait for `open` to be called.
-      await new Promise((resolve) => void self.open.callsFake(resolve));
-
-      expect(self.open).to.have.been.calledOnce;
-
-      // Wait for timeout to complete.
-      clock.tick(100);
-
-      expect(self.open).to.have.been.calledTwice;
-    });
-  });
-
-  describe('gaaNotifySignIn', () => {
-    it('posts message when passed a user', () => {
-      self.opener = self;
-      sandbox.stub(self, 'postMessage');
-      const gaaUser = {
-        email: 'email',
-        familyName: 'familyName',
-        givenName: 'givenName',
-        idToken: 'idToken',
-        imageUrl: 'imageUrl',
-        name: 'name',
-        authorizationData: {
-          /* eslint-disable google-camelcase/google-camelcase */
-          access_token: 'accessToken',
-          id_token: 'idToken',
-          scope: 'scope',
-          expires_in: 0,
-          first_issued_at: 0,
-          expires_at: 0,
-          /* eslint-enable google-camelcase/google-camelcase */
-        },
-      };
-      GaaGoogle3pSignInButton.gaaNotifySignIn({gaaUser});
-
-      expect(self.postMessage).to.have.been.calledWithExactly({
-        stamp: POST_MESSAGE_STAMP,
-        command: POST_MESSAGE_COMMAND_USER,
-        gaaUser,
-      });
-    });
-  });
-});
 
 describes.realWin('GaaMetering', () => {
   let microdata;
